Lock package.json actions while writing in progress
I noticed that if I were to quickly start several actions that
write to package.hson (eg. deleting/updating dependencies), only
the first one would succeed. This is because it locks that file
while it's working, and so the other actions can't write to it.

I have some much better long-term ideas about how to handle this
complication internally, invisibly to the user, but for now I went
with the simple approach: Disable all actions that write to the
package.json until the current action completes.